feat: modular approach for extensibility and composability
